export siggraph schema as siggraph.rs
